+++ internal/resource/common/attributesources/attribute_sources_schema.go
@@ -17,12 +17,20 @@ import (
 	"github.com/pingidentity/terraform-provider-pingfederate/internal/resource/common/resourcelink"
 )
 
-func CommonAttributeSourceSchema(optionalAndComputedNestedAttributeContractFulfillment bool) map[string]schema.Attribute {
+func CommonAttributeSourceSchema(optionalAndComputedNestedAttributeContractFulfillment, includeDataStoreRefResourceLinkLocation bool) map[string]schema.Attribute {
 	commonAttributeSourceSchema := map[string]schema.Attribute{}
+
+	var resourceLinkAttributes map[string]schema.Attribute
+	if includeDataStoreRefResourceLinkLocation {
+		resourceLinkAttributes = resourcelink.ToSchema()
+	} else {
+		resourceLinkAttributes = resourcelink.ToSchemaNoLocation()
+	}
+
 	commonAttributeSourceSchema["data_store_ref"] = schema.SingleNestedAttribute{
 		Required:    true,
 		Description: "Reference to the associated data store.",
-		Attributes:  resourcelink.ToSchemaLocationUseStateForUnknown(),
+		Attributes:  resourceLinkAttributes,
 	}
 	commonAttributeSourceSchema["id"] = schema.StringAttribute{
 		Optional:    true,
@@ -40,8 +48,8 @@ func CommonAttributeSourceSchema(optionalAndComputedNestedAttributeContractFulfi
 	return commonAttributeSourceSchema
 }
 
-func CustomAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContractFulfillment bool) map[string]schema.Attribute {
-	customAttributeSourceSchema := CommonAttributeSourceSchema(optionalAndComputedNestedAttributeContractFulfillment)
+func CustomAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContractFulfillment, includeDataStoreRefResourceLinkLocation bool) map[string]schema.Attribute {
+	customAttributeSourceSchema := CommonAttributeSourceSchema(optionalAndComputedNestedAttributeContractFulfillment, includeDataStoreRefResourceLinkLocation)
 	customAttributeSourceSchema["type"] = schema.StringAttribute{
 		Computed:    true,
 		Optional:    false,
@@ -67,8 +75,8 @@ func CustomAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeCon
 	return customAttributeSourceSchema
 }
 
-func JdbcAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContractFulfillment bool) map[string]schema.Attribute {
-	jdbcAttributeSourceSchema := CommonAttributeSourceSchema(optionalAndComputedNestedAttributeContractFulfillment)
+func JdbcAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContractFulfillment, includeDataStoreRefResourceLinkLocation bool) map[string]schema.Attribute {
+	jdbcAttributeSourceSchema := CommonAttributeSourceSchema(optionalAndComputedNestedAttributeContractFulfillment, includeDataStoreRefResourceLinkLocation)
 	jdbcAttributeSourceSchema["type"] = schema.StringAttribute{
 		Computed:    true,
 		Optional:    false,
@@ -95,8 +103,8 @@ func JdbcAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContr
 	return jdbcAttributeSourceSchema
 }
 
-func LdapAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContractFulfillment bool) map[string]schema.Attribute {
-	ldapAttributeSourceSchema := CommonAttributeSourceSchema(optionalAndComputedNestedAttributeContractFulfillment)
+func LdapAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContractFulfillment, includeDataStoreRefResourceLinkLocation bool) map[string]schema.Attribute {
+	ldapAttributeSourceSchema := CommonAttributeSourceSchema(optionalAndComputedNestedAttributeContractFulfillment, includeDataStoreRefResourceLinkLocation)
 	ldapAttributeSourceSchema["type"] = schema.StringAttribute{
 		Required:    true,
 		Description: "The data store type of this attribute source.",
@@ -151,8 +159,8 @@ func LdapAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContr
 	return ldapAttributeSourceSchema
 }
 
-func ToSchema(sizeAtLeast int, optionalAndComputedNestedAttributeContractFulfillment bool) schema.ListNestedAttribute {
-	attributeSourcesDefault, _ := types.ListValue(types.ObjectType{AttrTypes: ElemAttrType()}, nil)
+func ToSchema(sizeAtLeast int, optionalAndComputedNestedAttributeContractFulfillment, includeDataStoreRefResourceLinkLocation bool) schema.ListNestedAttribute {
+	attributeSourcesDefault, _ := types.ListValue(types.ObjectType{AttrTypes: ElemAttrType(includeDataStoreRefResourceLinkLocation)}, nil)
 	validators := []validator.List{}
 	if sizeAtLeast > 0 {
 		validators = append(validators, listvalidator.SizeAtLeast(sizeAtLeast))
@@ -168,7 +176,7 @@ func ToSchema(sizeAtLeast int, optionalAndComputedNestedAttributeContractFulfill
 				"custom_attribute_source": schema.SingleNestedAttribute{
 					Description: "The configured settings used to look up attributes from a custom data store.",
 					Optional:    true,
-					Attributes:  CustomAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContractFulfillment),
+					Attributes:  CustomAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContractFulfillment, includeDataStoreRefResourceLinkLocation),
 					Validators: []validator.Object{
 						objectvalidator.ExactlyOneOf(
 							path.MatchRelative().AtParent().AtName("jdbc_attribute_source"),
@@ -179,7 +187,7 @@ func ToSchema(sizeAtLeast int, optionalAndComputedNestedAttributeContractFulfill
 				"jdbc_attribute_source": schema.SingleNestedAttribute{
 					Description: "The configured settings used to look up attributes from a JDBC data store.",
 					Optional:    true,
-					Attributes:  JdbcAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContractFulfillment),
+					Attributes:  JdbcAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContractFulfillment, includeDataStoreRefResourceLinkLocation),
 					Validators: []validator.Object{
 						objectvalidator.ExactlyOneOf(
 							path.MatchRelative().AtParent().AtName("custom_attribute_source"),
@@ -190,7 +198,7 @@ func ToSchema(sizeAtLeast int, optionalAndComputedNestedAttributeContractFulfill
 				"ldap_attribute_source": schema.SingleNestedAttribute{
 					Description: "The configured settings used to look up attributes from a LDAP data store.",
 					Optional:    true,
-					Attributes:  LdapAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContractFulfillment),
+					Attributes:  LdapAttributeSourceSchemaAttributes(optionalAndComputedNestedAttributeContractFulfillment, includeDataStoreRefResourceLinkLocation),
 					Validators: []validator.Object{
 						objectvalidator.ExactlyOneOf(
 							path.MatchRelative().AtParent().AtName("custom_attribute_source"),