+++ internal/resource/common/resourcelink/resource_link_schema.go
@@ -2,6 +2,8 @@ package resourcelink
 
 import (
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 )
 
 func ToSchemaLocationUseStateForUnknown() map[string]schema.Attribute {
@@ -10,10 +12,34 @@ func ToSchemaLocationUseStateForUnknown() map[string]schema.Attribute {
 			Description: "The ID of the resource.",
 			Required:    true,
 		},
+		"location": schema.StringAttribute{
+			DeprecationMessage: "This field is now deprecated and will be removed in a future release.",
+			Description:        "A read-only URL that references the resource. If the resource is not currently URL-accessible, this property will be null.",
+			Computed:           true,
+			Optional:           false,
+			PlanModifiers: []planmodifier.String{
+				stringplanmodifier.UseStateForUnknown(),
+			},
+		},
 	}
 }
 
 func ToSchema() map[string]schema.Attribute {
+	return map[string]schema.Attribute{
+		"id": schema.StringAttribute{
+			Description: "The ID of the resource.",
+			Required:    true,
+		},
+		"location": schema.StringAttribute{
+			DeprecationMessage: "This field is now deprecated and will be removed in a future release.",
+			Description:        "A read-only URL that references the resource. If the resource is not currently URL-accessible, this property will be null.",
+			Computed:           true,
+			Optional:           false,
+		},
+	}
+}
+
+func ToSchemaNoLocation() map[string]schema.Attribute {
 	return map[string]schema.Attribute{
 		"id": schema.StringAttribute{
 			Description: "The ID of the resource.",