+++ internal/resource/config/oauth/tokenexchange/tokengeneratormapping/oauth_token_exchange_token_generator_mapping_resource.go
@@ -52,7 +52,7 @@ func (r *oauthTokenExchangeTokenGeneratorMappingResource) Schema(ctx context.Con
 	schema := schema.Schema{
 		Description: "Manages the mapping from a token exchange processor policy to a token generator.",
 		Attributes: map[string]schema.Attribute{
-			"attribute_sources":              attributesources.ToSchema(0, false),
+			"attribute_sources":              attributesources.ToSchema(0, false, true),
 			"attribute_contract_fulfillment": attributecontractfulfillment.ToSchema(true, false, false),
 			"issuance_criteria":              issuancecriteria.ToSchema(),
 			"source_id": schema.StringAttribute{
@@ -115,7 +115,7 @@ func (r *oauthTokenExchangeTokenGeneratorMappingResource) Configure(_ context.Co
 
 func readOauthTokenExchangeTokenGeneratorMappingResourceResponse(ctx context.Context, r *client.ProcessorPolicyToGeneratorMapping, state *oauthTokenExchangeTokenGeneratorMappingResourceModel) diag.Diagnostics {
 	var diags, respDiags diag.Diagnostics
-	state.AttributeSources, respDiags = attributesources.ToState(ctx, r.AttributeSources)
+	state.AttributeSources, respDiags = attributesources.ToState(ctx, r.AttributeSources, true)
 	diags.Append(respDiags...)
 	state.AttributeContractFulfillment, respDiags = attributecontractfulfillment.ToState(ctx, r.AttributeContractFulfillment)
 	diags.Append(respDiags...)